# src/app/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "",
    "description": "",
    "version": "1.0"
  },
  "produces": ["application/json"],
  "paths": {
...
    "/houses/{id}/tasks": {
      "get": {
        "operationId": "fetchHouseTasks",
        "tags": ["/houses"],
        "produces": [
          "application/json"
        ],
        "summary": "",
        "description": "Fetch joined tasks of house",
        "parameters": [{
          "name": "id",
          "in": "path",
          "description": "ID of house that needs to be fetched belonging tasks",
          "required": true,
          "type": "integer",
          "minimum": 1,
          "format": "int32"
        }],
        "responses": {
          "200": {
            "description": "House tasks successfully fetched",
            "schema": {
              "$ref": "#/definitions/Task"
            }
          }
        },
        "security": [{ "JWT": [] }]
      }
    },
...
    "/users/{id}/tasks": {
      "get": {
        "operationId": "fetchUserTasks",
        "tags": ["/users"],
        "produces": [
          "application/json"
        ],
        "summary": "",
        "description": "Fetch joined tasks of user",
        "parameters": [{
          "name": "id",
          "in": "path",
          "description": "ID of user that needs to be fetched belonging tasks",
          "required": true,
          "type": "integer",
          "minimum": 1,
          "format": "int32"
        }],
        "responses": {
          "200": {
            "description": "User tasks successfully fetched",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        },
        "security": [{ "JWT": [] }]
      }
    },
    "/tasks": {
      "get": {
        "operationId": "fetchTasks",
        "tags": ["/tasks"],
        "produces": [
          "application/json"
        ],
        "summary": "",
        "description": "Fetch tasks",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Tasks successfully fetched",
            "schema": {
              "$ref": "#/definitions/Task"
            }
          }
        },
        "security": [{ "JWT": [] }]
      },
      "post": {
        "operationId": "createTask",
        "tags": ["/tasks"],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "summary": "",
        "description": "Create new task",
        "parameters": [{
          "in": "body",
          "name": "body",
          "description": "Create new task with description.",
          "required": true,
          "schema": {
            "$ref": "#/definitions/NewTask"
          }
        }],
        "responses": {
          "200": {
            "description": "New task successfully created.",
            "schema": {
              "$ref": "#/definitions/Task"
            }
          }
        },
        "security": [{ "JWT": [] }]
      }
    },
    "/tasks/{id}": {
      "get": {
        "operationId": "fetchTaskById",
        "tags": ["/tasks"],
        "produces": [
          "application/json"
        ],
        "summary": "",
        "description": "Fetch task with id",
        "parameters": [{
          "name": "id",
          "in": "path",
          "description": "ID of task that needs to be fetched",
          "required": true,
          "type": "integer",
          "minimum": 1,
          "format": "int32"
        }],
        "responses": {
          "200": {
            "description": "Task successfully fetched",
            "schema": {
              "$ref": "#/definitions/Task"
            }
          }
        },
        "security": [{ "JWT": [] }]
      },
      "put": {
        "operationId": "updateTask",
        "tags": ["/tasks"],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "summary": "",
        "description": "Update existed task",
        "parameters": [{
          "name": "id",
          "in": "path",
          "description": "ID of task that needs to be updated",
          "required": true,
          "type": "integer",
          "minimum": 1,
          "format": "int32"
        }],
        "responses": {
          "200": {
            "description": "Task successfully updated.",
            "schema": {
              "$ref": "#/definitions/Task"
            }
          }
        },
        "security": [{ "JWT": [] }]
      },
      "delete": {
        "operationId": "deleteTask",
        "tags": ["/tasks"],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "summary": "",
        "description": "Delete existed task",
        "parameters": [{
          "name": "id",
          "in": "path",
          "description": "ID of task that needs to be deleted",
          "required": true,
          "type": "integer",
          "minimum": 1,
          "format": "int32"
        }],
        "responses": {
          "200": {
            "description": "Task successfully deleted."
          }
        },
        "security": [{ "JWT": [] }]
      }
    }
  },
  "securityDefinitions": {
    "JWT": {
      "description": "",
      "type": "apiKey",
      "name": "authorization",
      "in": "header"
    }
  },
  "definitions": {
...
    "User": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "username": {
          "type": "string"
        },
        "houseId": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
    "Task": {
      "type": "object",
      "required": [],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "description": {
          "type": "string"
        },
        "done": {
          "type": "boolean"
        },
        "userId": {
          "type": "integer",
          "format": "int32"
        },
        "houseId": {
          "type": "integer",
          "format": "int32"
        },
        "createdUserId": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "NewTask": {
      "type": "object",
      "required": [
        "description",
        "houseId"
      ],
      "properties": {
        "description": {
          "type": "string"
        },
        "houseId": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
  }
}